Instantiate process only at begining of clock-tick
Created a new prefab for process while on hold which waits for next clock-tick to start then instantiates the process then self-destructs.

Works with timer paused. Note that if you pause at timer=0, the process is created even though the timer is paused.

Processes are still only spawned with hard-coded instruction list.